html/semantics/embedded-content-0/media-elements/canPlayType_supported_but_no_codecs_parameter_1.html

<!doctype html>
<html>
 <head>
  <title>{audio,video}.canPlayType() - supported format w/o codecs parameter</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-navigator-canplaytype">spec reference</a></p>
  <audio id="a">
  </audio>
  <video id="v">
  </video>
  <div id="log"></div>
  <script>
test(function() {
 var v = document.getElementById("a");
 if (v.canPlayType('audio/ogg; codecs="vorbis"') != "") {
  assert_equals(
   v.canPlayType("audio/ogg"),
   "maybe",
   "audioElement.canPlayType() should return 'maybe' on supported format w/o codecs parameter");
 }
}, "audio.canPlayType() - supported format w/o codecs parameter");

test(function() {
 var v = document.getElementById("v");
 if (v.canPlayType('video/webm; codecs="vp8, vorbis"') != "") {
   assert_equals(
    v.canPlayType("video/webm"),
    "maybe",
    "videoElement.canPlayType() should return 'maybe' on supported format w/o codecs parameter");
 }
}, "video.canPlayType() - supported format w/o codecs parameter");
  </script>
 </body>
</html>
